+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/filecache/FiberSensor.scala
@@ -25,8 +25,8 @@ import com.google.common.base.Throwables
 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.execution.datasources.oap.IndexMeta
-import org.apache.spark.sql.execution.datasources.oap.listener.SparkListenerCustomInfoUpdate
 import org.apache.spark.sql.execution.datasources.oap.utils.CacheStatusSerDe
+import org.apache.spark.sql.oap.listener.SparkListenerCustomInfoUpdate
 import org.apache.spark.util.collection.BitSet
 
 private[oap] case class FiberCacheStatus(
