+++ src/main/scala/org/apache/spark/sql/hive/thriftserver/OapEnv.scala
@@ -22,9 +22,9 @@ import java.io.PrintStream
 import org.apache.spark.{SparkConf, SparkContext}
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SQLContext
-import org.apache.spark.sql.execution.datasources.oap.listener.OapListener
 import org.apache.spark.sql.hive.{HiveUtils, OapSessionState}
 import org.apache.spark.sql.oap.OapSession
+import org.apache.spark.sql.oap.listener.OapListener
 import org.apache.spark.sql.oap.ui.OapTab
 import org.apache.spark.util.Utils
